SVG: parse internal svg elements, fallback size
Parse all svg elements and apply viewBox transformations for each.
Elements without a viewBox are supported and this fixes an issue
with rendering SVGs that do not define a top-level viewBox.

If no width, height, or viewBox is specified on the top-level root
then decode to a default image size of 300x150 pixels.